add CRUD methods

+++ src/main/java/com/endava/internship/dataCenter/controller/EmployeesController.java
@@ -90,15 +90,51 @@ public class EmployeesController {
 
     // PUT - update an existing employee
     @PutMapping("/employees/{id}")
-    public Employees updateEmployeeById(@PathVariable Integer id, @RequestBody EmployeeDto employeeDto){
-        return employeesService.updateEmployees(id, employeeDto);
+    public ResponseEntity<String> updateEmployeeById(@PathVariable Integer id, @RequestBody EmployeeDto employeeDto){
+        //  First name and Last name - not to be null, empty or blank
+        String checkFirstName = employeeDto.getFirstName();
+        String checkLastName = employeeDto.getLastName();
+        if(checkFirstName == null && checkLastName == null && checkFirstName.trim().isEmpty() && checkLastName.trim().isEmpty()){
+            return new ResponseEntity<>("First name and Last name - not to be null, empty or blank", HttpStatus.BAD_REQUEST);
+        }
+
+        // Email - to match email format
+        boolean checkEmail = employeeDto.getEmail().matches("[a-zA-Z0-9\\+\\.\\_\\%\\-\\+]{1,256}" + "\\@" + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,64}" + "("
+                + "\\." + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,25}" + ")+");
+        if(!checkEmail){
+            return new ResponseEntity<>("Email is not correct", HttpStatus.BAD_REQUEST);
+        }
+
+        // Phone number - to start with 0 and and contain exactly 9 digits
+        boolean checkPhoneNumber = employeeDto.getPhoneNumber().matches("^0\\d{9}");
+        if(!checkPhoneNumber){
+            return new ResponseEntity<>("Phone number is not correct. Need to start with 0 and contain exactly 9 digits", HttpStatus.BAD_REQUEST);
+        }
+
+        // Salary - min 1.0
+        boolean checkSalary = employeeDto.getSalary() >= 1.0;
+        if(!checkSalary){
+            return new ResponseEntity<>("Check salary. Min value is 1.0", HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>("Employee updated", HttpStatus.OK);
     }
 
 
     // DELETE  -  delete employee by id
     @DeleteMapping("/employees/{id}")
-    public void deleteEmployee(@PathVariable Integer id){
+    public ResponseEntity<String> deleteEmployee(@PathVariable Integer id){
+        if(id == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         employeesService.deleteEmployee(id);
+
+        Employees employee = this.employeesService.getEmployeeById(id);
+
+        if(employee == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<>("Employee deleted", HttpStatus.OK);
     }
 
 }
